+++ spec/features/staff/fund_managers_can_view_funds_spec.rb
@@ -11,7 +11,7 @@ RSpec.feature "Fund managers can view funds on an organisation page" do
 
   context "when the user is a fund_manager" do
     before do
-      authenticate!(user: create(:fund_manager, organisations: []))
+      authenticate!(user: create(:fund_manager))
     end
 
     scenario "the user will see them on the organisation show page" do
@@ -35,7 +35,7 @@ RSpec.feature "Fund managers can view funds on an organisation page" do
 
   context "when the user is a delivery_partner" do
     before do
-      authenticate!(user: create(:delivery_partner, organisations: [organisation]))
+      authenticate!(user: create(:delivery_partner, organisation: organisation))
     end
 
     scenario "the user will not see them on the show page for their organisation" do