+++ spec/features/staff/users_can_view_a_transaction_as_xml_spec.rb
@@ -7,7 +7,7 @@ RSpec.feature "Users can view a transaction as XML" do
   let(:fund) { create(:fund, organisation: organisation) }
   let!(:activity) { create(:activity, hierarchy: fund) }
   let(:transaction) { create(:transaction, fund: fund) }
-  let(:user) { create(:administrator, organisations: [organisation]) }
+  let(:user) { create(:administrator, organisation: organisation) }
 
   context "when the user is not logged in" do
     it "redirects the user to the root path" do